+++ manifest/protobuf.go
@@ -22,10 +22,19 @@ func loadProtobufs(pkg *pbsubstreams.Package, manif *Manifest) error {
 		seen[*file.Name] = true
 	}
 
-	var importPaths []string
+	importPaths := []string{}
 	for _, imp := range manif.Protobuf.ImportPaths {
 		importPaths = append(importPaths, manif.resolvePath(imp))
 	}
+
+	// The manifest's root directory is always added to the list of import paths so that
+	// files specified relative to the manifest's directory works properly. It is added last
+	// so that if user's specified import paths contains the file, it's picked from their
+	// import paths instead of the implicitely added folder.
+	if manif.Workdir != "" {
+		importPaths = append(importPaths, manif.Workdir)
+	}
+
 	// User-specified protos
 	parser := protoparse.Parser{
 		ImportPaths:           importPaths,
@@ -34,9 +43,10 @@ func loadProtobufs(pkg *pbsubstreams.Package, manif *Manifest) error {
 
 	for _, file := range manif.Protobuf.Files {
 		if seen[file] {
-			return fmt.Errorf("WARNING: proto file %s already exists in system protobufs, do not include in your manifest", file)
+			return fmt.Errorf("WARNING: proto file %s already exists in system protobufs, do not include it in your manifest", file)
 		}
 	}
+
 	customFiles, err := parser.ParseFiles(manif.Protobuf.Files...)
 	if err != nil {
 		return fmt.Errorf("error parsing proto files %q (import paths: %q): %w", manif.Protobuf.Files, importPaths, err)